show user solves from the leaderboard
