+++ src/routes/user/[userId]/+page.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+import ProblemDisplay from '$lib/components/ProblemDisplay.svelte';
+
+// The userId will be extracted from the URL in the ProblemDisplay component
+</script>
+
+<ProblemDisplay pageTitle="User Problems" defaultSolvedFilterState="solved" />
