+++ src/routes/user/[userId]/+page.ts
@@ -1,0 +1,7 @@
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = ({ params }) => {
+  return {
+    userId: params.userId
+  };
+};
